requirements.txt
typer[all]
rich
radon
pycodestyle
bandit
pytest
coverage
safety
gitpython
pyinstrument
pytest-benchmark
memory_profiler
scipy
numpy
google-generativeai>=0.3.1
ruff
mypy
pytest-cov
pyyaml
lizard 
python-dotenv
litellm>=1.35.0
fastapi
starlette
pydantic
httpx
anyio
requests